+++ src/RoadRegistry.Hosts/Metadata/IMetadataUpdater.cs
@@ -1,0 +1,10 @@
+namespace RoadRegistry.Hosts.Metadata
+{
+    using System.Threading;
+    using System.Threading.Tasks;
+
+    public interface IMetadataUpdater
+    {
+        Task UpdateAsync(CancellationToken cancellationToken);
+    }
+}