+++ src/RoadRegistry.Product.Schema/MigrationTables.cs
@@ -1,6 +1,0 @@
-namespace RoadRegistry.Product.Schema;
-
-public static class MigrationTables
-{
-    public const string Product = "__EFMigrationsHistoryProduct";
-}