+++ src/RoadRegistry.Projector/Projections/Response/SyndicationStatus.cs
@@ -1,0 +1,9 @@
+namespace RoadRegistry.Projector.Projections.Response
+{
+    public class SyndicationStatus
+    {
+        public string ProjectionName { get; set; }
+
+        public long Position { get; set; }
+    }
+}