+++ src/RoadRegistry.Wfs.Projections/RoadSegmentRecordProjection.cs
@@ -74,7 +74,7 @@ namespace RoadRegistry.Wfs.Projections
                     Geometry2D = WfsGeometryTranslator.Translate2D(envelope.Message.Geometry),
                     MorphologyDutchName = morphology.Translation.Name,
                     StatusDutchName = status.Translation.Name,
-                    AccessRestrictionId = accessRestriction.Translation.Identifier,
+                    AccessRestriction = accessRestriction.Translation.Name,
                     LeftSideStreetNameId = envelope.Message.LeftSide.StreetNameId,
                     LeftSideStreetName = leftSideStreetNameRecord?.DutchNameWithHomonymAddition ??
                                          envelope.Message.LeftSide.StreetName,
@@ -142,7 +142,7 @@ namespace RoadRegistry.Wfs.Projections
                 Geometry2D = WfsGeometryTranslator.Translate2D(roadSegmentAdded.Geometry),
                 MorphologyDutchName = morphology.Translation.Name,
                 StatusDutchName = status.Translation.Name,
-                AccessRestrictionId = accessRestriction.Translation.Identifier,
+                AccessRestriction = accessRestriction.Translation.Name,
                 LeftSideStreetNameId = roadSegmentAdded.LeftSide.StreetNameId,
                 LeftSideStreetName = leftSideStreetNameRecord?.DutchName,
                 RightSideStreetNameId = roadSegmentAdded.RightSide.StreetNameId,
@@ -191,7 +191,7 @@ namespace RoadRegistry.Wfs.Projections
             roadSegmentRecord.Geometry2D = WfsGeometryTranslator.Translate2D(roadSegmentModified.Geometry);
             roadSegmentRecord.MorphologyDutchName = morphology.Translation.Name;
             roadSegmentRecord.StatusDutchName = status.Translation.Name;
-            roadSegmentRecord.AccessRestrictionId = accessRestriction.Translation.Identifier;
+            roadSegmentRecord.AccessRestriction = accessRestriction.Translation.Name;
             roadSegmentRecord.LeftSideStreetNameId = roadSegmentModified.LeftSide.StreetNameId;
             roadSegmentRecord.LeftSideStreetName = leftSideStreetNameRecord?.DutchName;
             roadSegmentRecord.RightSideStreetNameId = roadSegmentModified.RightSide.StreetNameId;