+++ src/RoadRegistry.Wms.ProjectionHost/Metadata/IMetadataUpdater.cs
@@ -1,9 +1,0 @@
-namespace RoadRegistry.Wms.ProjectionHost.Metadata;
-
-using System.Threading;
-using System.Threading.Tasks;
-
-public interface IMetadataUpdater
-{
-    Task UpdateAsync(CancellationToken cancellationToken);
-}