+++ test/RoadRegistry.Wms.ProjectionHost.Tests/Projections/Framework/ExpectedWegsegmentRecord.cs
@@ -4,7 +4,7 @@ using System;
 
 public class ExpectedWegsegmentRecord
 {
-    public int? wegsegmentID { get; set; }
+    public int wegsegmentID { get; set; }
     public int? methode { get; set; }
     public string beheerder { get; set; }
     public DateTime? begintijd { get; set; }